% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tech_shares.R
\name{calc_iea_tech_shares}
\alias{calc_iea_tech_shares}
\title{Calculates industry energy shares for IEA data}
\usage{
calc_iea_tech_shares(
  psut_df,
  U_matname = "U",
  V_matname = "V",
  input_fractions_column = "U_fractions",
  output_fractions_column = "V_fractions"
)
}
\arguments{
\item{psut_df}{A data frame of PSUT matrices}

\item{U_matname}{The name of the \strong{U} matrix in \code{psut_df}.
Default is "U".}

\item{V_matname}{The name of the \strong{V} matrix in \code{psut_df}.
Default is "V".}

\item{input_fractions_column}{The name of the output column that contains vectors of input fractions.}

\item{output_fractions_column}{The name of the output column that contains vectors of output fractions.}
}
\value{
A version of \code{psut_df} with input and output industry shares added as column vectors at the right side of the data frame.
}
\description{
Calculates both for input (from the \strong{U} matrix)
and for outputs (from the \strong{V} matrix).
}
