% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{create_iea_eta_i_reports}
\alias{create_iea_eta_i_reports}
\title{Create a series of efficiency reports}
\usage{
create_iea_eta_i_reports(eta_i_df, reports_dest_folder, release = FALSE)
}
\arguments{
\item{eta_i_df}{A data frame of machine efficiencies in vectors.}

\item{reports_dest_folder}{The output folder for these reports.}

\item{release}{A boolean that tells whether to write the reports.
Default is \code{FALSE}.}
}
\value{
Nothing. This function should be called for its side effect of
creating reports.
}
\description{
Creates reports of efficiencies for machines in PSUT data frames.
Each report contains efficiencies for a single machine
across all countries.
}
\details{
The report is focused IEA data where the last stage is final energy.
}
