% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tech_shares.R
\name{expand_tech_shares}
\alias{expand_tech_shares}
\title{Expand a data frame of industry shares}
\usage{
expand_tech_shares(
  tech_shares_vectors,
  input_fractions_column = "U_fractions",
  output_fractions_column = "V_fractions"
)
}
\arguments{
\item{tech_shares_vectors}{A data frame produced by \code{calc_iea_industry_shares()}.}

\item{input_fractions_column}{The name of the output column that contains vectors of input fractions.}

\item{output_fractions_column}{The name of the output column that contains vectors of output fractions.}
}
\value{
A data frame with \code{tech_shares_vectors} expanded to tidy format.
}
\description{
Expand a data frame of industry shares
}
