% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposition_pipelines.R
\name{get_pipeline}
\alias{get_pipeline}
\title{Create a targets workflow for decomposition work}
\usage{
get_pipeline(
  countries = "all",
  years = "all",
  database_version,
  targeted_aggregations_file,
  pipeline_releases_folder,
  pipeline_caches_folder,
  reports_dest_folder,
  release = FALSE
)
}
\arguments{
\item{countries}{A string vector of 3-letter country codes.
Default is "all", meaning all available countries should be analyzed.}

\item{years}{A numeric vector of years to be analyzed.
Default is "all", meaning all available years should be analyzed.}

\item{database_version}{The version of the database we'll use.
See details.}

\item{targeted_aggregations_file}{The Excel file that describes industry aggregations
(in an "industry_aggregations" tab) and product aggregations
(in a "product_aggregations" tab).
Both tabs should contain
a Many column, a Few column,
a Country column, and a Year column.
The string "All" in Country or Year
designates that all countries or years should be aggregated.}

\item{pipeline_releases_folder}{The path to a folder where releases of output targets are pinned.}

\item{pipeline_caches_folder}{The path to a folder where releases of pipeline caches are stored.}

\item{reports_dest_folder}{The destination folder for reports.}

\item{release}{Boolean that tells whether to do a release of the results.
Default is \code{FALSE}.}
}
\value{
A list of \code{tar_target}s to be executed in a workflow.
}
\description{
This is a target factory whose arguments
specify the details of a targets workflow to be constructed.
}
