% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregations.R
\name{load_agg_map}
\alias{load_agg_map}
\title{Load an aggregation map}
\usage{
load_agg_map(
  aggregation_file,
  aggregation_tab = tools::file_path_sans_ext(basename(aggregation_file)),
  country_colname = "Country",
  year_colname = "Year",
  many_colname = "Many",
  few_colname = "Few"
)
}
\arguments{
\item{aggregation_file}{The path to the industry aggregations file.}

\item{aggregation_tab}{The tab in \code{industry_aggregations_file} that contains
an industry aggregation table.
Default is \code{aggregation_file |> basename() |> tools::file_path_sans_ext()}}

\item{country_colname}{The name of the country column in \code{aggregation_tab}.
Default is "Country".}

\item{year_colname}{The name of the year column in \code{aggregation_tab}.
Default is "Year".}

\item{many_colname}{The name of the many column on the \code{industry_aggregations_tab}.
Default is "Many".}

\item{few_colname}{The name of the few column on the \code{industry_aggregations_tab}.
Default is "Few".}
}
\value{
An aggregation map.
}
\description{
An aggregation file is an Excel file that
contains an \code{aggregation_tab} with
\code{many_colname} and \code{few_colname},
by default "Many" and "Few".
This function loads the aggregation table
and converts to an aggregation map
using \code{matsbyname::agg_table_to_agg_map()}.
}
