% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregations.R
\name{psut_aggregation}
\alias{psut_aggregation}
\title{Perform an aggregation according to an aggregation map}
\usage{
psut_aggregation(
  psut_df,
  aggregation_map,
  country_colname = "Country",
  year_colname = "Year",
  agg_map_colname = "agg_map",
  margin = c("Product", "Industry"),
  matcols = c("R", "U", "U_feed", "U_EIOU", "r_EIOU", "V", "Y", "S_units"),
  matnames_colname = "matnames",
  matvals_colname = "matvals"
)
}
\arguments{
\item{psut_df}{A data frame of PSUT matrices.}

\item{aggregation_map}{An aggregation map.}

\item{rowcoltype}{The row or column types to be aggregated.
Default is \code{c("Product", "Industry")}.}
}
\value{
A version of \code{psut_df} with aggregations according to \code{aggregation_map} and
\code{rowcoltype}.
}
\description{
Perform an aggregation according to an aggregation map
}
