% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregations.R
\name{targeted_aggregation}
\alias{targeted_aggregation}
\title{Targeted product and industry aggregation according to an aggregation map}
\usage{
targeted_aggregation(
  psut_df,
  aggregation_map,
  country_colname = "Country",
  year_colname = "Year",
  agg_map_colname = "agg_map",
  margin = c("Industry", "Product"),
  matcols = c("R", "U", "U_feed", "U_EIOU", "r_EIOU", "V", "Y", "S_units"),
  matnames_colname = "matnames",
  matvals_colname = "matvals"
)
}
\arguments{
\item{psut_df}{A data frame of PSUT matrices with \code{country_colname}, \code{year_colname},
and columns of PSUT matrices (\code{matcols}).}

\item{aggregation_map}{An aggregation map with \code{country_colname}, \code{year_colname}, and}

\item{rowcoltype}{The row or column types to be aggregated.
Default is \code{c("Product", "Industry")}.}
}
\value{
A version of \code{psut_df} with aggregations according to \code{aggregation_map} and
\code{rowcoltype}.
}
\description{
\code{matsbyname::aggregate_byname()} performs aggregation of
rows and columns but is unaware of countries and years.
This function performs targeted aggregation
for some or all countries and years
by including a \code{country_colname} and a \code{year_colname}
in the \code{aggregation_map}.
In addition,
the Country and Year columns
may contain optional "All" strings
to indicate all countries or years should be aggregated
as specified.
}
