% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tech_shares.R
\name{calc_iea_industry_shares}
\alias{calc_iea_industry_shares}
\title{Calculates industry energy shares for IEA data}
\usage{
calc_iea_industry_shares(
  psut_df,
  R_matname = Recca::psut_cols$R,
  U_matname = Recca::psut_cols$U,
  U_feed_matname = Recca::psut_cols$U_feed,
  U_eiou_matname = Recca::psut_cols$U_eiou,
  r_eiou_matname = Recca::psut_cols$r_eiou,
  V_matname = Recca::psut_cols$V,
  Y_matname = Recca::psut_cols$Y,
  S_units_matname = Recca::psut_cols$S_units,
  Y_fu_details_matname = Recca::psut_cols$Y_fu_details,
  U_eiou_fu_details_matname = Recca::psut_cols$U_eiou_fu_details,
  input_fractions_column = "U_fractions",
  output_fractions_column = "V_fractions"
)
}
\arguments{
\item{psut_df}{A data frame of PSUT matrices}

\item{R_matname, U_feed_matname, U_eiou_matname, r_eiou_matname, V_matname, Y_matname, S_units_matname, Y_fu_details_matname, U_eiou_matname_matname}{Columns deleted on output.}

\item{U_matname}{The name of the \strong{U} matrix in \code{psut_df}.
This column is deleted on output.
Default is "U".}

\item{V_matname}{The name of the \strong{V} matrix in \code{psut_df}.
This column is deleted on output.
Default is "V".}

\item{input_fractions_column}{The name of the output column that contains vectors of input fractions.}

\item{output_fractions_column}{The name of the output column that contains vectors of output fractions.}
}
\value{
A version of \code{psut_df} with input and output industry shares added as column vectors at the right side of the data frame.
}
\description{
Calculates both for input (from the \strong{U} matrix)
and for outputs (from the \strong{V} matrix).
All incoming PSUT matrix columns in \code{psut_df} are dropped.
The output is two coumns of column vectors.
}
